+++ src/models/layerMetadata/layer3DMetadata.ts
@@ -1,4 +1,5 @@
 import { GeoJSON } from 'geojson';
+import { RecordType, ProductType, RecordStatus } from '@map-colonies/types';
 import { graphql } from '../common/decorators/graphQL/graphql.decorator';
 import {
   FieldCategory,
@@ -7,12 +8,10 @@ import {
   IFieldConfigInfo,
   IPropFieldConfigInfo,
 } from '../common/decorators/fieldConfig/fieldConfig.decorator';
-import { RecordType } from '../pycsw/coreEnums';
+import { IMetadataCommonModel } from './interfaces/metadataCommonModel';
 import { getPyCSWMapping, IPYCSWMapping, pycsw } from './decorators/property/csw.decorator';
 import { tsTypes, TsTypes } from './decorators/property/tsTypes.decorator';
-import { ProductType, RecordStatus } from './enums';
 import { catalogDB, ORMColumnType } from './decorators/property/catalogDB.decorator';
-import { IMetadataCommonModel } from './interfaces/metadataCommonModel';
 
 export interface ILayer3DMetadata {
   srsId: string | undefined;