+++ src/models/layerMetadata/layerDEMMetadata.ts
@@ -1,4 +1,5 @@
 import { GeoJSON } from 'geojson';
+import { RecordType, ProductType, RecordStatus } from '@map-colonies/types';
 import { IPropCatalogDBMapping } from '../common/interfaces/propCatalogDBMapping.interface';
 import { graphql } from '../common/decorators/graphQL/graphql.decorator';
 import {
@@ -8,12 +9,11 @@ import {
   IFieldConfigInfo,
   IPropFieldConfigInfo,
 } from '../common/decorators/fieldConfig/fieldConfig.decorator';
-import { DEMDataType, NoDataValue, RecordType, UndulationModel } from '../pycsw/coreEnums';
+import { DEMDataType, NoDataValue, UndulationModel } from '../pycsw/coreEnums';
 import { IMetadataCommonModel } from './interfaces/metadataCommonModel';
 import { getPyCSWMapping, IPYCSWMapping, pycsw } from './decorators/property/csw.decorator';
 import { getCatalogDBMapping, ICatalogDBMapping, catalogDB } from './decorators/property/catalogDB.decorator';
 import { getTsTypesMapping, tsTypes, TsTypes } from './decorators/property/tsTypes.decorator';
-import { ProductType, RecordStatus } from './enums';
 
 export interface ILayerMetadata {
   srsId: string | undefined;