+++ src/models/raster/ingestion/zod/schemas/index.ts
@@ -1,3 +1,3 @@
-export * from './partData.schema';
 export * from './inputFiles.schema';
 export * from './metadata.schema';
+export * from './polygonParts.schema';