+++ src/models/rasterCatalog/rasterCatalogUpsertRequest.ts
@@ -1,7 +1,18 @@
 import { Link } from '../common';
-import { LayerMetadata } from '../layerMetadata';
+import { LayerMetadata, PycswLayerCatalogRecord } from '../layerMetadata';
+
+export type EditLayerMetadata = Partial<
+  Pick<
+    PycswLayerCatalogRecord,
+    'classification' | 'productName' | 'productSubType' | 'description' | 'producerName' | 'region' | 'scale' | 'keywords'
+  >
+>;
 
 export interface IRasterCatalogUpsertRequestBody {
   metadata: LayerMetadata;
   links?: Link[];
+}
+
+export interface IRasterCatalogEditRequestBody {
+  metadata: EditLayerMetadata;
 }